Replace SequentialRangeSampler and ShuffledRangeSampler with a simplified OffsetSampler.
